Rename OIDC auth extension
