+++ extension/oidcauthextension/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package authoidcextension
+package oidcauthextension
 
 import "go.opentelemetry.io/collector/config"
